+++ src/services/movies.services.ts
@@ -1,5 +1,5 @@
 import { db } from '../database'
-import { createMovieQuery, getMovieByTitleQuery } from '../database/sqlScipt'
+import { createMovieQuery, getMovieByTitleQuery, markMovieAsWatchedQuery } from '../database/sqlScipt'
 import { MovieWithCategory } from '../types'
 
 // GET movies services
@@ -14,6 +14,7 @@ export const getMovieByTitle = async (title: string): Promise<MovieWithCategory>
     throw new Error('Error connecting to db')
   }
 }
+
 /* export const getMovies */
 
 // CREATE movies services
@@ -29,5 +30,15 @@ export const createMovie = async (title: string, description: string, releaseDat
 }
 
 // UPDATE movies services
+// i will put the following service here cause i don't fin a better place to
+export const markMovieAsWatched = async (userId: string, movieId: string): Promise<void> => {
+  const query = markMovieAsWatchedQuery
+  try {
+    await db.query(query, [userId, movieId])
+  } catch (err: any) {
+    console.log(err.message)
+    throw new Error('Error connecting to db')
+  }
+}
 
 // DELETE movies services